benches: Use Criterion.rs instead of Divan
Divan [1] is simpler, has nicer output, and seems to be more actively
maintained than Criterion.rs [2], which is why I originally used it to
implement the benchmarks. However, I decided to switch to Criterion.rs
now.

The reason is that Criterion.rs implements better statistical analysis
and allows to easily compare benchmark runs, including saving and using
baselines. For me, this turned out to be a killer feature.

[1] https://crates.io/crates/divan
[2] https://crates.io/crates/criterion