scanner: Make `Scanner` not implement `Iterator`
Until now, `Scanner` implemented `Iterator`. This was nice in principle,
and allowed us to use `std::iter::Peekable`, making the parser quite
elegant. But it also meant the `next` method returned `Option<Token>`,
with the end of input signaled by returning `None`. Handling this at
multiple places in the parser was really annoying.

As a solution, this commit makes `Scanner` not implement `Iterator`. The
`next` method now returns just `Token`, with the end of input signaled
by a new `TokenKind::Eof` token. Peeking is implemented manually for the
time being (the plan is to rewrite this part in a different way later).