parser: Rewrite the parser
Until now, the parser was based on `peek` and `next` operations, which
operated on the next unconsumed token. While this was conceptually
simple and elegant, these operations were fairly low-level, didn't help
with producing error messages, and required tracking of the peeked
token.

This commit rewrites the parser in terms of more high-level `accept_any`
and `expect` operations, which are themselves implemented in terms of
more low-level `current` and `advance` operations (which are also used
in the parser alongside the high-level ones). At the same time, the
rewrite introduces a notion of the current token, replacing tracking of
the peeked token. The overall result is simpler code and more extensible
(and more traditional) parser architecture.

Note the rewrite was made possible by a previous change that made token
kind a first-class value.