Implement position information tracking
We now track position information for all tokens, AST nodes, and errors.
This information is used for error reporting.

The tokens and AST nodes store their span, with its ends represented by
a character offset. The errors also store their span, but with its ends
represented by a character offset, line, and column. Translation between
these two representations is done using a source map filled-in by the
scanner.

This architecture means we don't waste space by storing line and column
for each token and AST node, but we have them available in errors. The
cost is some amount of complexity.

For a simple arithmetic expression evaluator, all this is a bit of an
overkill, but I wanted to explore techniques used in production parsers,
compilers, and similar tools.